Move checkDataIntegrity to test-helpers.
