+++ src/selectors/index.ts
@@ -5,7 +5,6 @@ export { default as attribute } from './attribute'
 export { default as attributeEquals } from './attributeEquals'
 export { default as canShowModal } from './canShowModal'
 export { default as chain } from './chain'
-export { default as checkDataIntegrity } from './checkDataIntegrity'
 export { default as decodeContextUrl } from './decodeContextUrl'
 export { default as decodeThoughtsUrl } from './decodeThoughtsUrl'
 export { default as exists } from './exists'
